backend/requirements.txt
deepgram-sdk==3.2.0
groq==0.4.2
websockets==12.0
pyaudio==0.2.14
pyaudiowpatch==0.2.12.5; sys_platform == 'win32'
python-dotenv==1.0.0
